+++ torchbit/__init__.py
@@ -1,3 +1,8 @@
+import cocotb
+from packaging import version
+
+assert version.parse(cocotb.__version__) >= version.parse("2.0.0"), "Cocotb version must be == 2.x for torchbit >= 2.x, please upgrade cocotb."
+
 from . import core
 from . import debug
 from . import tools